MULE-15562: Fix Mule-4-Status-Windows pipeline
- Regex property must be set to false otherwise the settings file path
passed by Maven's property `mule.extension.archetype.testSettings` will
be treated as a regex, therefore on Windows Systems the backslash will
be treated as a special char and will be deleted leaving a corrupted
file path.

- By default, this property is set to true on replacer's plugin config.
When so, the replacer plugin uses under the hood the
java.util.regex.Matcher#replaceAll method in its TokenReplacer class
where the actual replacement is carried on.

- According to Java docs "Note that backslashes (\) and dollar signs ($)
in the replacement string may cause the results to be different than if
it were being treated as a literal replacement string. [...] and
backslashes are used to escape literal characters in the replacement
string."